TutorialManager 구조 변경
다른 Manager에서 관리하던 튜토리얼 함수들을 TutorialManager 및 DiaryManager로 이동했습니다.
